fix: simplify DataTable to the bare bones of Observable's implementation
